Create Snake game using python
This repository contains a  Snake Game implemented using Python and the Turtle graphics library. The game features a blue game screen with a black boundary, a snake that grows longer as it eats food, and keeps track of the current and high scores. The game speeds up as the snake grows, adding an extra challenge for players.